[zprezto-update] Add convenience function to update zprezto (#1344)
* [zprezto-update] Add convenience function to update zprezto

This function checks if there is any update to zprezto, and if so
will pull in the changes. It will not attempt a pull unless
it is fastforwardable. It also makes sure the user is on the master
branch before attempting.

* [zprezto-update] Improve resilience of the function

Better error checking of status of the git repository
and better error producing.

Fit columns into mostly 80 width and add a missing printf
argument.

Use ( ) around the function so changing directory does not affect
the outer scope.

* [README] Add instructions on using zprezto-update function